修复设置窗口无法自动保存的bug

+++ UI_WPF/Pages/Settings/Setting_CustomPage.xaml.cs
@@ -89,7 +89,7 @@ namespace Another_Mirai_Native.UI.Pages
                                 switch (item.Value.Type)
                                 {
                                     case JTokenType.Integer:
-                                        success = int.TryParse(textBox.Text, out int intValue);
+                                        success = long.TryParse(textBox.Text, out long intValue);
                                         saveValue = intValue;
                                         break;
